Inline getIsAnchorFmSignup() utility function
